+++ R/propose_snapshot.R
@@ -18,7 +18,6 @@
 #'   to download those packages. Both these files are written to the
 #'   directory given by the `path_staging` argument.
 #' @inheritParams update_staging
-#' @param repo_staging Character string, URL of the staging universe.
 #' @param types Character vector, what to pass to the `types` field in the
 #'   snapshot API URL. Controls the types of binaries and documentation
 #'   included in the snapshot.
@@ -27,58 +26,45 @@
 #' url_staging = "https://github.com/r-multiverse/staging"
 #' path_staging <- tempfile()
 #' gert::git_clone(url = url_staging, path = path_staging)
-#' propose_snapshot(
-#'   path_staging = path_staging,
-#'   repo_staging = "https://staging.r-multiverse.org"
-#' )
+#' propose_snapshot(path_staging = path_staging)
 #' }
 propose_snapshot <- function(
   path_staging,
-  repo_staging = "https://staging.r-multiverse.org",
-  types = c("src", "win", "mac"),
-  mock = NULL
+  types = c("src", "win", "mac")
 ) {
-  path_issues <- file.path(path_staging, "issues.json")
-  issues <- character(0L)
-  if (file.exists(path_issues)) {
-    issues <- names(jsonlite::read_json(path_issues, simplifyVector = TRUE))
+  path_freeze <- file.path(path_staging, "freeze.json")
+  freeze <- character(0L)
+  if (file.exists(path_freeze)) {
+    freeze <- jsonlite::read_json(path_freeze, simplifyVector = TRUE)
   }
   file_staging <- file.path(path_staging, "packages.json")
   json_staging <- jsonlite::read_json(file_staging, simplifyVector = TRUE)
-  json_staging <- json_staging[, c("package", "url", "branch")]
-  meta_staging <- mock$staging %||% meta_packages(repo_staging)
-  meta_staging <- meta_staging[, c("package", "remotesha")]
-  staging <- merge(
-    x = json_staging,
-    y = meta_staging,
-    all.x = FALSE,
-    all.y = FALSE
-  )
-  staging <- staging[staging$branch == staging$remotesha,, drop = FALSE] # nolint
-  staging <- staging[!(staging$package %in% issues),, drop = FALSE] # nolint
-  staging$remotesha <- NULL
+  exclude <- setdiff(json_staging$package, freeze)
+  json_staging <- json_staging[json_staging$package %in% freeze,, drop = FALSE] # nolint
   file_snapshot <- file.path(path_staging, "snapshot.json")
-  jsonlite::write_json(staging, file_snapshot, pretty = TRUE)
-  r_version <- staging_r_version()
+  jsonlite::write_json(json_staging, file_snapshot, pretty = TRUE)
+  r_version <- r_version_staging()
   binaries <- paste0("&binaries=", r_version$short)
   url <- paste0(
-    "https://staging.r-multiverse.org/api/snapshot/zip",
+    "https://staging.r-multiverse.org/api/snapshot/tar",
     "?types=",
     paste(types, collapse = ","),
     binaries,
-    "&packages=",
-    paste(staging$package, collapse = ",")
+    "&skip_packages=",
+    paste(exclude, collapse = ",")
   )
   writeLines(url, file.path(path_staging, "snapshot.url"))
-  writeLines(r_version$full, file.path(path_staging, "r_version_full.txt"))
-  writeLines(r_version$short, file.path(path_staging, "r_version_short.txt"))
-  writeLines(
-    staging_start(),
-    file.path(path_staging, "date_staging_start.txt")
+  meta <- list(
+    date = data.frame(
+      staging = date_staging(),
+      snapshot = date_snapshot()
+    ),
+    r_version = r_version[c("full", "short")]
   )
-  writeLines(
-    as.character(Sys.Date()),
-    file.path(path_staging, "date_snapshot.txt")
+  jsonlite::write_json(
+    meta,
+    file.path(path_staging, "meta.json"),
+    pretty = TRUE
   )
   invisible()
 }